refactor tests for mysql & sqlite
