+++ yaroo/lib/db/db.dart
@@ -11,7 +11,7 @@ class UseDatabaseConnection {
 
   UseDatabaseConnection(this.name) : _driver = DB.driver(name);
 
-  Query query(String table) => Query.query(table).driver(_driver);
+  Query query(String table) => Query.table(table).driver(_driver);
 }
 
 class DB {
